Text overflow bug fixes in the posts
