+++ src/pages/Timeline/index.jsx
@@ -7,7 +7,7 @@ import UserContext from '../../contexts/UserContext';
 import Header from '../shared/Header';
 import CreatePost from './CreatePost';
 import CirclesLoader from '../shared/CirclesLoader';
-import HashtagTrending from '../shared/HashtagTrending';
+import Trending from '../shared/Trending';
 import NoPostMessage from '../shared/NoPostMessage';
 
 export default function Timeline(){
@@ -52,7 +52,7 @@ export default function Timeline(){
 					}
 				</TimelineContent>
 				<HashtagContainer>
-					<HashtagTrending/>
+					<Trending/>
 				</HashtagContainer>
 			</Background>
 		</>
